feature : 상품 단건조회, 수정, 삭제 기능 구현

+++ src/main/java/com/test08/domain/product/repository/ProductRepository.java
@@ -3,5 +3,7 @@ package com.test08.domain.product.repository;
 import com.test08.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+    // productId int라 Integer로 변경
+
 }
